+++ apps/io-wallet-user-func/src/infra/http/handlers/__test__/create-wallet-attestation-v2.spec.ts
@@ -1,14 +1,15 @@
+/* eslint-disable max-lines-per-function */
 import { AttestationService } from "@/attestation-service";
 import { iOSMockData } from "@/infra/mobile-attestation-service/ios/__test__/config";
 import { NonceRepository } from "@/nonce";
 import { WalletInstanceRepository } from "@/wallet-instance";
-import { GRANT_TYPE_KEY_ATTESTATION } from "@/wallet-provider";
 import * as H from "@pagopa/handler-kit";
 import * as L from "@pagopa/logger";
 import { FiscalCode, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import { UrlFromString } from "@pagopa/ts-commons/lib/url";
 import * as appInsights from "applicationinsights";
 import { decode } from "cbor-x";
+import * as crypto from "crypto";
 import * as E from "fp-ts/Either";
 import * as O from "fp-ts/Option";
 import * as TE from "fp-ts/TaskEither";
@@ -16,7 +17,10 @@ import { flow } from "fp-ts/lib/function";
 import * as jose from "jose";
 import { describe, expect, it } from "vitest";
 
-import { CreateWalletAttestationV2Handler } from "../create-wallet-attestation-v2";
+import {
+  CreateWalletAttestationV2Handler,
+  WalletAttestations,
+} from "../create-wallet-attestation-v2";
 import { privateEcKey, publicEcKey, signer } from "./keys";
 
 const { assertion, challenge, hardwareKey, keyId } = iOSMockData;
@@ -40,16 +44,18 @@ const url = flow(
   }),
 );
 
-const entityConfiguration = {
-  authorityHints: [url("https://ta.example.org")],
-  federationEntity: {
-    basePath: url("https://wallet-provider.example.org"),
-    homepageUri: url("https://wallet-provider.example.org/privacy_policy"),
-    logoUri: url("https://wallet-provider.example.org/logo.svg"),
-    organizationName: "wallet provider" as NonEmptyString,
-    policyUri: url("https://wallet-provider.example.org/info_policy"),
-    tosUri: url("https://wallet-provider.example.org/logo.svg"),
-  },
+const federationEntity = {
+  basePath: url("https://wallet-provider.example.org"),
+  homepageUri: url("https://wallet-provider.example.org/privacy_policy"),
+  logoUri: url("https://wallet-provider.example.org/logo.svg"),
+  organizationName: "wallet provider" as NonEmptyString,
+  policyUri: url("https://wallet-provider.example.org/info_policy"),
+  tosUri: url("https://wallet-provider.example.org/logo.svg"),
+};
+
+const walletAttestationClaims = {
+  walletLink: "https://foo.com",
+  walletName: "Wallet name",
 };
 
 const mockAttestationService: AttestationService = {
@@ -101,10 +107,14 @@ const telemetryClient: appInsights.TelemetryClient = {
   trackException: () => void 0,
 } as unknown as appInsights.TelemetryClient;
 
+function isStringArray(u: unknown): u is string[] {
+  return Array.isArray(u) && u.every((item) => typeof item === "string");
+}
+
 describe("CreateWalletAttestationV2Handler", async () => {
   const josePrivateKey = await jose.importJWK(privateEcKey);
   const walletAttestationRequest = await new jose.SignJWT({
-    challenge,
+    aud: "aud",
     cnf: {
       jwk: publicEcKey,
     },
@@ -112,45 +122,50 @@ describe("CreateWalletAttestationV2Handler", async () => {
     hardware_signature: signature.toString("base64"),
     integrity_assertion: authenticatorData.toString("base64"),
     iss: "demokey",
+    nonce: challenge, // TODO: rename challenge in nonce
     sub: "https://wallet-provider.example.org/",
   })
     .setProtectedHeader({
       alg: "ES256",
       kid: publicEcKey.kid,
-      typ: "war+jwt",
+      typ: "wp-war+jwt",
     })
     .setIssuedAt()
     .setExpirationTime("2h")
     .sign(josePrivateKey);
 
+  const req = {
+    ...H.request("https://wallet-provider.example.org"),
+    body: {
+      assertion: walletAttestationRequest,
+    },
+    method: "POST",
+  };
+
   it("should return a 200 HTTP response on success", async () => {
-    const req = {
-      ...H.request("https://wallet-provider.example.org"),
-      body: {
-        assertion: walletAttestationRequest,
-        grant_type: GRANT_TYPE_KEY_ATTESTATION,
-      },
-      method: "POST",
-    };
     const handler = CreateWalletAttestationV2Handler({
       attestationService: mockAttestationService,
-      entityConfiguration,
+      federationEntity,
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
       nonceRepository,
       signer,
       telemetryClient,
+      walletAttestationConfig: walletAttestationClaims,
       walletInstanceRepository,
     });
 
-    const result = await handler();
-    expect.assertions(3);
-    expect(result).toEqual({
+    await expect(handler()).resolves.toEqual({
       _tag: "Right",
       right: {
         body: expect.objectContaining({
-          wallet_attestation: expect.any(String),
+          wallet_attestations: expect.arrayContaining([
+            expect.objectContaining({
+              format: expect.any(String),
+              wallet_attestation: expect.any(String),
+            }),
+          ]),
         }),
         headers: expect.objectContaining({
           "Content-Type": "application/json",
@@ -158,15 +173,129 @@ describe("CreateWalletAttestationV2Handler", async () => {
         statusCode: 200,
       },
     });
+  });
 
-    // check trailing slashes are removed
+  it("should return a correctly encoded jwt on success and URLs within the token should not have trailing slashes", async () => {
+    const handler = CreateWalletAttestationV2Handler({
+      attestationService: mockAttestationService,
+      federationEntity,
+      input: req,
+      inputDecoder: H.HttpRequest,
+      logger,
+      nonceRepository,
+      signer,
+      telemetryClient,
+      walletAttestationConfig: walletAttestationClaims,
+      walletInstanceRepository,
+    });
+
+    const result = await handler();
+    expect.assertions(4);
+
     if (E.isRight(result)) {
-      const body = result.right.body;
-      const walletAttestation = body.wallet_attestation;
-      if (typeof walletAttestation === "string") {
-        const decoded = jose.decodeJwt(walletAttestation);
-        expect((decoded.iss || "").endsWith("/")).toBe(false);
-        expect((decoded.sub || "").endsWith("/")).toBe(false);
+      const body = WalletAttestations.decode(result.right.body);
+      if (E.isRight(body)) {
+        const walletAttestations = body.right.wallet_attestations;
+        const walletAttestationJwt = walletAttestations.find(
+          (walletAttestation) => walletAttestation.format === "jwt",
+        );
+        if (walletAttestationJwt) {
+          const jwtHeader = jose.decodeProtectedHeader(
+            walletAttestationJwt.wallet_attestation,
+          );
+          expect(Object.keys(jwtHeader).sort()).toEqual(
+            ["alg", "typ", "kid", "trust_chain"].sort(),
+          );
+          const jwtPayload = jose.decodeJwt(
+            walletAttestationJwt.wallet_attestation,
+          );
+          expect(Object.keys(jwtPayload).sort()).toEqual(
+            [
+              "aal",
+              "exp",
+              "iat",
+              "cnf",
+              "iss",
+              "sub",
+              "wallet_name",
+              "wallet_link",
+            ].sort(),
+          );
+          // check trailing slashes are removed
+          expect((jwtPayload.iss || "").endsWith("/")).toBe(false);
+          expect((jwtPayload.sub || "").endsWith("/")).toBe(false);
+        }
+      }
+    }
+  });
+
+  it("should return a correctly encoded sdjwt with disclosures on success and URLs within the token should not have trailing slashes", async () => {
+    const handler = CreateWalletAttestationV2Handler({
+      attestationService: mockAttestationService,
+      federationEntity,
+      input: req,
+      inputDecoder: H.HttpRequest,
+      logger,
+      nonceRepository,
+      signer,
+      telemetryClient,
+      walletAttestationConfig: walletAttestationClaims,
+      walletInstanceRepository,
+    });
+
+    const result = await handler();
+    expect.assertions(6);
+
+    if (E.isRight(result)) {
+      const body = WalletAttestations.decode(result.right.body);
+      if (E.isRight(body)) {
+        const walletAttestations = body.right.wallet_attestations;
+        const walletAttestationDcSdJwt = walletAttestations.find(
+          (walletAttestation) => walletAttestation.format === "dc+sd-jwt",
+        );
+        if (walletAttestationDcSdJwt) {
+          const splittedVpToken =
+            walletAttestationDcSdJwt.wallet_attestation.split("~");
+          const sdJwt = splittedVpToken[0];
+          const disclosures = splittedVpToken.slice(1);
+
+          // check the properties of the header and payload
+          const jwtHeader = jose.decodeProtectedHeader(sdJwt);
+          expect(Object.keys(jwtHeader).sort()).toEqual(
+            ["alg", "typ", "kid", "trust_chain"].sort(),
+          );
+          const jwtPayload = jose.decodeJwt(sdJwt);
+          expect(Object.keys(jwtPayload).sort()).toEqual(
+            [
+              "aal",
+              "exp",
+              "iat",
+              "cnf",
+              "iss",
+              "sub",
+              "vct",
+              "_sd",
+              "sd_alg",
+            ].sort(),
+          );
+
+          // check that the hashed disclosures are included in _sd
+          const _sd = jwtPayload._sd;
+          if (isStringArray(_sd)) {
+            disclosures.forEach((disclosure) => {
+              const disclosureDigest = crypto
+                .createHash("sha256")
+                .update(Buffer.from(disclosure, "base64"))
+                .digest("base64");
+
+              expect(_sd.includes(disclosureDigest)).toBe(true);
+            });
+          }
+
+          // check trailing slashes are removed
+          expect((jwtPayload.iss || "").endsWith("/")).toBe(false);
+          expect((jwtPayload.sub || "").endsWith("/")).toBe(false);
+        }
       }
     }
   });
@@ -175,20 +304,20 @@ describe("CreateWalletAttestationV2Handler", async () => {
     const req = {
       ...H.request("https://wallet-provider.example.org"),
       body: {
-        assertion: walletAttestationRequest,
-        grant_type: "foo",
+        assertion1: "foo",
       },
       method: "POST",
     };
     const handler = CreateWalletAttestationV2Handler({
       attestationService: mockAttestationService,
-      entityConfiguration,
+      federationEntity,
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
       nonceRepository,
       signer,
       telemetryClient,
+      walletAttestationConfig: walletAttestationClaims,
       walletInstanceRepository,
     });
 
@@ -223,19 +352,19 @@ describe("CreateWalletAttestationV2Handler", async () => {
       ...H.request("https://wallet-provider.example.org"),
       body: {
         assertion: walletAttestationRequest,
-        grant_type: GRANT_TYPE_KEY_ATTESTATION,
       },
       method: "POST",
     };
     const handler = CreateWalletAttestationV2Handler({
       attestationService: mockAttestationService,
-      entityConfiguration,
+      federationEntity,
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
       nonceRepository,
       signer,
       telemetryClient,
+      walletAttestationConfig: walletAttestationClaims,
       walletInstanceRepository: walletInstanceRepositoryWithRevokedWI,
     });
 
@@ -265,19 +394,19 @@ describe("CreateWalletAttestationV2Handler", async () => {
       ...H.request("https://wallet-provider.example.org"),
       body: {
         assertion: walletAttestationRequest,
-        grant_type: GRANT_TYPE_KEY_ATTESTATION,
       },
       method: "POST",
     };
     const handler = CreateWalletAttestationV2Handler({
       attestationService: mockAttestationService,
-      entityConfiguration,
+      federationEntity,
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
       nonceRepository,
       signer,
       telemetryClient,
+      walletAttestationConfig: walletAttestationClaims,
       walletInstanceRepository: walletInstanceRepositoryWithNotFoundWI,
     });